organizando login menu
